scraper now working
